requirements.txt
tensorflow
pyyaml
pandas
matplotlib
